Evaluation code with attention map
